[MIG] hr_attendance_report_theoretical_time: Migration to 17.0
