+++ hr_attendance_report_theoretical_time/README.rst
@@ -149,6 +149,7 @@ Contributors
    -  Pedro M. Baeza.
    -  David Vidal
    -  Víctor Martínez
+   -  Juan José Seguí
 
 -  Pedro Gonzalez <pedro.gonzalez@pesol.es>
 -  Aritz Olea <ao@landoo.es>
