+++ hr_attendance_report_theoretical_time/readme/CONTRIBUTORS.md
@@ -2,6 +2,7 @@
   - Pedro M. Baeza.
   - David Vidal
   - Víctor Martínez
+  - Juan José Seguí
 - Pedro Gonzalez \<<pedro.gonzalez@pesol.es>\>
 - Aritz Olea \<<ao@landoo.es>\>
 - [Trobz](https://trobz.com):
