+++ hr_attendance_report_theoretical_time/reports/__init__.py
@@ -1,4 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from . import hr_attendance_report
 from . import hr_attendance_theoretical_time_report
