+++ hr_attendance_report_theoretical_time/security/hr_attendance_report_theoretical_time_security.xml
@@ -5,7 +5,10 @@
     <record model="ir.rule" id="rule_theoretical_vs_worked_report_own">
         <field name="name">Theoretical vs worked hours: Own attendances</field>
         <field name="model_id" ref="model_hr_attendance_theoretical_time_report" />
-        <field name="groups" eval="[(4, ref('hr_attendance.group_hr_attendance'))]" />
+        <field
+            name="groups"
+            eval="[(4, ref('hr_attendance.group_hr_attendance_own_reader'))]"
+        />
         <field name="domain_force">[['employee_id.user_id', '=', user.id]]</field>
     </record>
     <record model="ir.rule" id="rule_theoretical_vs_worked_report_all">
@@ -13,7 +16,7 @@
         <field name="model_id" ref="model_hr_attendance_theoretical_time_report" />
         <field
             name="groups"
-            eval="[(4, ref('hr_attendance.group_hr_attendance_user'))]"
+            eval="[(4, ref('hr_attendance.group_hr_attendance_officer'))]"
         />
         <field name="domain_force">[[1, '=', 1]]</field>
     </record>
