+++ hr_attendance_report_theoretical_time/security/ir.model.access.csv
@@ -1,4 +1,4 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_hr_attendance_theoretical_time_report,access_hr_attendance_theoretical_time_report,model_hr_attendance_theoretical_time_report,hr_attendance.group_hr_attendance,1,0,0,0
-access_wizard_theoretical_time,access_wizard_theoretical_time,model_wizard_theoretical_time,hr_attendance.group_hr_attendance_user,1,1,1,1
+access_hr_attendance_theoretical_time_report,access_hr_attendance_theoretical_time_report,model_hr_attendance_theoretical_time_report,hr_attendance.group_hr_attendance_own_reader,1,0,0,0
+access_wizard_theoretical_time,access_wizard_theoretical_time,model_wizard_theoretical_time,hr_attendance.group_hr_attendance_officer,1,1,1,1
 access_recompute_theoretical_attendance,access_recompute_theoretical_attendance,model_recompute_theoretical_attendance,hr_attendance.group_hr_attendance_manager,1,1,1,1
