+++ hr_attendance_report_theoretical_time/static/description/index.html
@@ -496,6 +496,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <li>Pedro M. Baeza.</li>
 <li>David Vidal</li>
 <li>Víctor Martínez</li>
+<li>Juan José Seguí</li>
 </ul>
 </li>
 <li>Pedro Gonzalez &lt;<a class="reference external" href="mailto:pedro.gonzalez&#64;pesol.es">pedro.gonzalez&#64;pesol.es</a>&gt;</li>
