+++ hr_attendance_report_theoretical_time/wizards/wizard_theoretical_time.xml
@@ -68,7 +68,7 @@
         name="Select Employees"
         action="wizard_theoretical_time_act_window"
         parent="menu_hr_attendance_theoretical_root"
-        groups="hr_attendance.group_hr_attendance_user"
+        groups="hr_attendance.group_hr_attendance_officer"
         sequence="25"
     />
 </odoo>
